Bug 654 - Bill paid prior to delivery is shown as due in 'COD Settlement' page via 'Home Delivery'

+++ wepbase/src/main/java/com/wep/common/app/Database/ComplimentaryBillDetail.java
@@ -1,64 +1,64 @@
 /***************************************************************************
  * Project Name		:	VAJRA
- * 
+ *
  * File Name		:	ComplimentaryBillDetail
- * 
+ *
  * Purpose			:	Represents ComplimentaryBillDetail table, takes care of 
  * 						initializing assigning and returning values of all the variables.
- * 
+ *
  * DateOfCreation	:	06-February-2013
- * 
+ *
  * Author			:	Balasubramanya Bharadwaj B S
- * 
+ *
  ***************************************************************************/
 package com.wep.common.app.Database;
 
 public class ComplimentaryBillDetail {
-	
+
 	// Private variables
 	String strComplimentaryReason;
 	int iBillNumber;
-	float fPaidAmount; 
-	
+	double fPaidAmount;
+
 	// Default Constructor
 	public ComplimentaryBillDetail(){
 		this.iBillNumber = 0;
 		this.fPaidAmount = 0;
 		this.strComplimentaryReason = "";
 	}
-	
+
 	// Parameterized Constructor
-	public ComplimentaryBillDetail(int BillNumber,float PaidAmount,String ComplimentaryReason){
+	public ComplimentaryBillDetail(int BillNumber,double PaidAmount,String ComplimentaryReason){
 		this.iBillNumber = BillNumber;
 		this.fPaidAmount = PaidAmount;
 		this.strComplimentaryReason = ComplimentaryReason;
 	}
-	
+
 	// getting BillNumber
 	public int getBillNumber(){
 		return this.iBillNumber;
 	}
-	
+
 	// getting PaidAmount
-	public float getPaidAmount(){
+	public double getPaidAmount(){
 		return this.fPaidAmount;
 	}
-	
+
 	// getting ComplimentaryReason
 	public String getComplimentaryReason(){
 		return this.strComplimentaryReason;
 	}
-	
+
 	// setting BillNumber
 	public void setBillNumber(int BillNumber){
 		this.iBillNumber = BillNumber;
 	}
-	
+
 	// setting PaidAmount
-	public void setPaidAmount(float PaidAmount){
+	public void setPaidAmount(double PaidAmount){
 		this.fPaidAmount = PaidAmount;
 	}
-	
+
 	// setting ComplimentaryReason
 	public void setComplimentaryReason(String ComplimentaryReason){
 		this.strComplimentaryReason = ComplimentaryReason;
